docs: fix PanelHeader name
Вынес LegacyPanelHeaderContent, чтобы правильно подтягивался PanelHeader
